Switches to AAB package format for Android
Updates the Android build process to use the AAB (Android App Bundle) package format instead of APK.

This change streamlines the deployment process and allows for more efficient delivery of app resources to users. It removes the project path resolution from the build script, as it's no longer needed with the updated configuration. Also configures the sample project to use .net9 sdk version.